chore: move and integrate ConfigureEvmFor
