+++ crates/node/builder/src/components/execute.rs
@@ -1,7 +1,7 @@
 //! EVM component for the node builder.
 use crate::{BuilderContext, FullNodeTypes};
-use reth_evm::execute::BlockExecutorProvider;
-use reth_node_api::{ConfigureEvm, HeaderTy, TxTy};
+use reth_evm::{execute::BlockExecutorProvider, ConfigureEvmFor};
+use reth_node_api::NodeTypes;
 use std::future::Future;
 
 /// A type that knows how to build the executor types.
@@ -9,7 +9,7 @@ pub trait ExecutorBuilder<Node: FullNodeTypes>: Send {
     /// The EVM config to use.
     ///
     /// This provides the node with the necessary configuration to configure an EVM.
-    type EVM: ConfigureEvm<Header = HeaderTy<Node::Types>, Transaction = TxTy<Node::Types>>;
+    type EVM: ConfigureEvmFor<<Node::Types as NodeTypes>::Primitives>;
 
     /// The type that knows how to execute blocks.
     type Executor: BlockExecutorProvider<
@@ -26,7 +26,7 @@ pub trait ExecutorBuilder<Node: FullNodeTypes>: Send {
 impl<Node, F, Fut, EVM, Executor> ExecutorBuilder<Node> for F
 where
     Node: FullNodeTypes,
-    EVM: ConfigureEvm<Header = HeaderTy<Node::Types>, Transaction = TxTy<Node::Types>>,
+    EVM: ConfigureEvmFor<<Node::Types as NodeTypes>::Primitives>,
     Executor:
         BlockExecutorProvider<Primitives = <Node::Types as reth_node_api::NodeTypes>::Primitives>,
     F: FnOnce(&BuilderContext<Node>) -> Fut + Send,
